# Add check hook for pgstat_end_function_usage in fmgr