--- a/.github/scripts/create_extension.sql
+++ b/.github/scripts/create_extension.sql
@@ -9,6 +9,7 @@ ALTER USER :user CREATEDB;
 ALTER SYSTEM SET babelfishpg_tsql.database_name = :db;
 ALTER SYSTEM SET babelfishpg_tsql.migration_mode = :'migration_mode';
 ALTER SYSTEM SET babelfishpg_tds.port = :tsql_port;
+ALTER SYSTEM SET track_functions = 'pl';
 
 \if :parallel_query_mode
     ALTER SYSTEM SET parallel_setup_cost = 0;