### Description

In babelfish there could be cases where a procedure rollbacks/deletes it self during execution.
For example 
```
BEGIN TRAN
CREATE PROCEDURE poc1 AS ROLLBACK;
EXEC p;
```
Procedure 'proc1' does not exists after its execution since the transaction in which it was created
was rolled back by the procedure itself.
Now while executing rollback inside such procedures, the pg_stat entry for this proc is also deleted.
When we return to fmgr to end the proc invocation, we will overwrite on this freed block of memory.
In future, when alloc tries to allocate this freed block if memory to some other process,
it may hit an assert since this block was corrupted.

To avoid this, we add an additional check before writing on the pg_stat entry post invocation.
Check involves simply looking for a local cache pg_stat entry for given key.

Note* PG does not need this check since such a case is not possible there.

#### Performance Effect
Use the simplest of procedures to avoid diluting the performance changes and get the worst case scenario.
Before & after numbers are taken as average of 4 runs.
```
CREATE PROCEDURE proc1 AS SELECT 1;
CREATE PROCEDURE proc2 @in INT AS SELECT @in + 1;
```
**CASE 1**
```
DECLARE @i INT = 0;
DECLARE @dt DATETIME = getdate();
WHILE (@i < 1000000)
BEGIN
    EXEC proc1;
    SET @i = @i + 1;
END;
DECLARE @duration_ms int;
SET @duration_ms = DATEDIFF(ms, @dt, getdate());
SELECT @duration_ms
GO
```
Before Fix:
_track_functions -->_ disabled: 29189 ms enabled: 29367.75 **Change: +0.62%**
After Fix:
_track_functions -->_ disabled: 28261 ms enabled: 28266 **Change: ~0%**


**CASE 2**
```
DECLARE @i INT = 0;
DECLARE @dt DATETIME = getdate();
WHILE (@in < 1000000)
BEGIN
    EXEC proc2 1;
    SET @i = @i + 1;
END;
DECLARE @duration_ms int;
SET @duration_ms = DATEDIFF(ms, @dt, getdate());
SELECT @duration_ms
GO
```
Before Fix:
_track_functions -->_ disabled: 34730.25 ms enabled: 34732.25 **Change: ~0%**
After Fix:
_track_functions -->_ disabled: 33918.75 ms enabled: 34075.5 **Change: +0.4%**

Tests show no visible change in performance.


#### Engine PR: https://github.com/babelfish-for-postgresql/postgresql_modified_for_babelfish/pull/305
#### Extension PR: https://github.com/babelfish-for-postgresql/babelfish_extensions/pull/2352

### Issues Resolved

[BABEL-4740]

### Sign Off

Signed-off-by: Tanzeel Khan <tzlkhan@amazon.com>

### Test Scenarios Covered ###

1. Enable track_function GUC for github actions otherwise execution does not even enter the pg_stat related code.
2. Different permutation and variation of example in description.
3. Existing test cases like TestTransactionSupportForProcedure & TestSimpleErrorsWithImplicitTran crashes without this fix when track function GUC is enabled.
4. Todo: Check how much the check hook deteriorates the performance.

### Check List
- [x] Commits are signed per the DCO using --signoff 

By submitting this pull request, I confirm that my contribution is under the terms of the Apache 2.0 and PostgreSQL licenses, and grant any person obtaining a copy of the contribution permission to relicense all or a portion of my contribution to the PostgreSQL License solely to contribute all or a portion of my contribution to the PostgreSQL open source project.

For more information on following Developer Certificate of Origin and signing off your commits, please check [here](https://github.com/babelfish-for-postgresql/babelfish_extensions/blob/main/CONTRIBUTING.md#developer-certificate-of-origin).